Decouple dump_map from write_to_file and move the latter to main.rs
